# src/openapi.yaml
openapi: 3.0.3
info:
  description: API for systems used by the employees of Zlagoda shops
  title: Zlagoda
  version: 1.0.0
servers:
- url: https://localhost:3333
...
paths:
  /api/employee:
    get:
      operationId: getEmployeeList
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCriteria'
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeListResponse'
          description: List of employees
      summary: Get list of employees
      tags:
      - employee
    put:
      operationId: createEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
          description: Id of new employee
      summary: Create a new employee
      tags:
      - employee
  /api/employee/{id}:
    delete:
      operationId: deleteEmployee
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether employee was deleted
      summary: Delete an employee by id
      tags:
      - employee
    get:
      operationId: getEmployeeById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
          description: Employee by id
      summary: Get employee by id
      tags:
      - employee
    post:
      operationId: updateEmployee
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether employee was updated
      summary: Update existing employee
      tags:
      - employee
...
  /api/product-archetype:
    get:
      operationId: getProductArchetypeList
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductArchetypeCriteria'
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductArchetypeListResponse'
          description: List of product archetypes
      summary: Get list of product archetypes
      tags:
      - product-archetype
    put:
      operationId: createProductArchetype
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductArchetypeView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
          description: Id of new product archetype
      summary: Create a new product archetype
      tags:
      - product-archetype
  /api/product-archetype/{id}:
    delete:
      operationId: deleteProductArchetype
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether product archetype was deleted
      summary: Delete a product archetype by id
      tags:
      - product-archetype
    get:
      operationId: getProductArchetypeById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductArchetype'
          description: Product archetype by id
      summary: Get product archetype by id
      tags:
      - product-archetype
    post:
      operationId: updateProductArchetype
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductArchetypeView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether product archetype was updated
      summary: Update existing product archetype
      tags:
      - product-archetype
  /api/product:
    get:
      operationId: getProductList
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCriteria'
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
          description: List of products
      summary: Get list of products
      tags:
      - product
    put:
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
          description: Id of new product
      summary: Create a new product
      tags:
      - product
  /api/product/{id}:
    delete:
      operationId: deleteProduct
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether product was deleted
      summary: Delete a product by id
      tags:
      - product
    get:
      operationId: getProductById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: product by id
      summary: Get product by id
      tags:
      - product
    post:
      operationId: updateProdact
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether product was updated
      summary: Update existing product
      tags:
      - product
...
  /api/customer-card:
    get:
      operationId: getCustomerCardList
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCardCriteria'
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerCard'
          description: List of customer cards
      summary: Get list of customer cards
      tags:
      - customer-card
    put:
      operationId: createCustomerCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCardView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: integer
          description: Id of new customer card
      summary: Create a new customer card
      tags:
      - customer-card
  /api/customer-card/{id}:
    delete:
      operationId: deleteCustomerCard
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether customer card was deleted
      summary: Delete a customer card by id
      tags:
      - customer-card
    get:
      operationId: getCustomerCardById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerCard'
          description: Customer card by id
      summary: Get customer card by id
      tags:
      - customer-card
    post:
      operationId: updateCustomerCard
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCardView'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Boolean whether customer card was updated
      summary: Update existing customer card
      tags:
      - customer-card
components:
  schemas:
    EmployeeView:
      example:
        firstName: firstName
        lastName: lastName
        zipCode: zipCode
        patronymic: patronymic
        role: null
        phoneNumber: phoneNumber
        city: city
        street: street
        workStartDate: 2000-01-23
        salary: 0
        birthDate: 2000-01-23
      properties:
        firstName:
          title: firstName
          type: string
        lastName:
          title: lastName
          type: string
        patronymic:
          title: patronymic
          type: string
        role:
          $ref: '#/components/schemas/EmployeeRole'
        salary:
          title: salary
          type: integer
        workStartDate:
          format: date
          title: workStartDate
          type: string
        birthDate:
          format: date
          title: birthDate
          type: string
        phoneNumber:
          title: phoneNumber
          type: string
        city:
          title: city
          type: string
        street:
          title: street
          type: string
        zipCode:
          title: zipCode
          type: string
      required:
      - birthDate
      - city
      - firstName
      - lastName
      - phoneNumber
      - role
      - salary
      - street
      - workStartDate
      - zipCode
      title: EmployeeView
      type: object
    Employee:
      allOf:
      - properties:
          id:
            type: integer
        required:
        - id
        type: object
      - $ref: '#/components/schemas/EmployeeView'
      example:
        firstName: firstName
        lastName: lastName
        zipCode: zipCode
        patronymic: patronymic
        role: null
        phoneNumber: phoneNumber
        city: city
        street: street
        workStartDate: 2000-01-23
        id: 6
        salary: 1
        birthDate: 2000-01-23
      title: Employee
    EmployeeRole:
      enum:
      - CASHIER
      - MANAGER
      title: EmployeeRole
      type: string
...
    ProductArchetypeView:
      example:
        name: name
        description: description
        category: 0
        manufacturer: manufacturer
      properties:
        name:
          title: name
          type: string
        category:
          title: category
          type: integer
        manufacturer:
          title: manufacturer
          type: string
        description:
          title: description
          type: string
      required:
      - category
      - description
      - manufacturer
      - name
      title: ProductArchetypeView
      type: object
    ProductArchetype:
      allOf:
      - properties:
          id:
            type: integer
        required:
        - id
        type: object
      - $ref: '#/components/schemas/ProductArchetypeView'
      example:
        name: name
        description: description
        id: 6
        category: 1
        manufacturer: manufacturer
      title: ProductArchetype
    ProductView:
      example:
        archetype: 0
        quantity: 1
        price: 6
        upc: upc
      properties:
        archetype:
          title: archetype
          type: integer
        upc:
          title: upc
          type: string
        price:
          title: price
          type: integer
        quantity:
          title: quantity
          type: integer
      required:
      - archetype
      - price
      - quantity
      - upc
      title: ProductView
      type: object
    Product:
      allOf:
      - properties:
          id:
            type: integer
          hasDiscount:
            type: boolean
        required:
        - hasDiscount
        - id
        type: object
      - $ref: '#/components/schemas/ProductView'
      example:
        archetype: 1
        quantity: 5
        hasDiscount: true
        price: 5
        upc: upc
        id: 6
      title: Product
...
    CustomerCardView:
      example:
        firstName: firstName
        lastName: lastName
        zipCode: zipCode
        patronymic: patronymic
        phoneNumber: phoneNumber
        discountPercent: 0
        city: city
        street: street
      properties:
        firstName:
          title: firstName
          type: string
        lastName:
          title: lastName
          type: string
        patronymic:
          title: patronymic
          type: string
        phoneNumber:
          title: phoneNumber
          type: string
        discountPercent:
          title: discountPercent
          type: integer
        city:
          title: city
          type: string
        street:
          title: street
          type: string
        zipCode:
          title: zipCode
          type: string
      required:
      - discountPercent
      - firstName
      - lastName
      - phoneNumber
      title: CustomerCardView
      type: object
    CustomerCard:
      allOf:
      - properties:
          id:
            type: integer
        required:
        - id
        type: object
      - $ref: '#/components/schemas/CustomerCardView'
      example:
        firstName: firstName
        lastName: lastName
        zipCode: zipCode
        patronymic: patronymic
        phoneNumber: phoneNumber
        discountPercent: 6
        city: city
        street: street
        id: 0
      title: CustomerCard
    EmployeeListResponse:
      example:
        totalCount: 0
        items:
        - firstName: firstName
          lastName: lastName
          zipCode: zipCode
          patronymic: patronymic
          role: null
          phoneNumber: phoneNumber
          city: city
          street: street
          workStartDate: 2000-01-23
          id: 6
          salary: 1
          birthDate: 2000-01-23
        - firstName: firstName
          lastName: lastName
          zipCode: zipCode
          patronymic: patronymic
          role: null
          phoneNumber: phoneNumber
          city: city
          street: street
          workStartDate: 2000-01-23
          id: 6
          salary: 1
          birthDate: 2000-01-23
      properties:
        totalCount:
          title: totalCount
          type: integer
        items:
          items:
            $ref: '#/components/schemas/Employee'
          title: items
          type: array
      required:
      - items
      - totalCount
      title: EmployeeListResponse
      type: object
...
    ProductArchetypeListResponse:
      example:
        totalCount: 0
        items:
        - name: name
          description: description
          id: 6
          category: 1
          manufacturer: manufacturer
        - name: name
          description: description
          id: 6
          category: 1
          manufacturer: manufacturer
      properties:
        totalCount:
          title: totalCount
          type: integer
        items:
          items:
            $ref: '#/components/schemas/ProductArchetype'
          title: items
          type: array
      required:
      - items
      - totalCount
      title: ProductArchetypeListResponse
      type: object
    ProductListResponse:
      example:
        totalCount: 0
        items:
        - archetype: 1
          quantity: 5
          hasDiscount: true
          price: 5
          upc: upc
          id: 6
        - archetype: 1
          quantity: 5
          hasDiscount: true
          price: 5
          upc: upc
          id: 6
      properties:
        totalCount:
          title: totalCount
          type: integer
        items:
          items:
            $ref: '#/components/schemas/Product'
          title: items
          type: array
      required:
      - items
      - totalCount
      title: ProductListResponse
      type: object
...
    EmployeeCriteria:
      allOf:
      - $ref: '#/components/schemas/BaseCriteria'
      - properties:
          role:
            $ref: '#/components/schemas/EmployeeRole'
          lastName:
            type: string
        type: object
      example:
        lastName: lastName
        role: null
      title: EmployeeCriteria
...
    ProductArchetypeCriteria:
      allOf:
      - $ref: '#/components/schemas/BaseCriteria'
      example: {}
      title: ProductArchetypeCriteria
    ProductCriteria:
      allOf:
      - $ref: '#/components/schemas/BaseCriteria'
      - properties:
          categoryIds:
            items:
              type: integer
            type: array
          upc:
            type: string
          name:
            type: string
          hasDiscount:
            type: boolean
          query:
            type: string
        type: object
      example:
        categoryIds:
        - 0
        - 0
        hasDiscount: true
        query: query
        name: name
        upc: upc
      title: ProductCriteria
...
    CustomerCardCriteria:
      allOf:
      - $ref: '#/components/schemas/BaseCriteria'
      - properties:
          lastName:
            type: string
          phoneNumber:
            type: string
          query:
            type: string
        type: object
      example:
        lastName: lastName
        phoneNumber: phoneNumber
        query: query
      title: CustomerCardCriteria
    BaseCriteria:
      allOf:
      - properties:
          offset:
            type: integer
          limit:
            type: integer
        type: object
      - $ref: '#/components/schemas/SortCriteria'
      title: BaseCriteria
    SortCriteria:
      properties:
        sortField:
          title: sortField
          type: string
        sortAscending:
          title: sortAscending
          type: boolean
      title: SortCriteria
      type: object